rename the filter_condition.rs to filter.rs
